feat(system): 用户分页查询及详情查询接口
- 新增 SysUserPageDto 分页查询请求DTO（继承PageQuery，支持username/realName/deptId/status筛选）
- 新增 SysUserVo 分页查询响应VO（含deptName/roleIds/roleNames，脱敏不返回passwordHash）
- 新增 UserRoleDto 用户角色联表查询DTO
- SysUserController 新增分页查询接口 GET /system/users，补全@Tag/@Operation注解
- SysDeptController 补全@Tag/@Operation注解
- ISysUserService/SysUserServiceImpl 新增page分页查询和getDetailById详情查询
- 分页/详情查询使用MPJ联表查询用户+部门，selectAs显式映射id→userId、isAdmin→Boolean
- fillRoles方法通过MPJ联表user_role+role一条SQL回填roleIds/roleNames
- SysUserRoleMapper从BaseMapper改为MPJBaseMapper支持联表查询

+++ erp-server/erp-common/src/main/java/com/qiheng/erp/common/result/PageResult.java
@@ -14,14 +14,14 @@ public class PageResult<T> implements Serializable {
     //数据
     private List<T> records;
     //总数据数
-    private long total;
+    private int total;
     //当前页码
-    private long pageNum;
+    private int pageNum;
     //每页数据数
-    private long pageSize;
+    private int pageSize;
 
 
-    public static <T> PageResult<T> of(List<T> records, long total, long pageNum, long pageSize) {
+    public static <T> PageResult<T> of(List<T> records, int total, int pageNum, int pageSize) {
         PageResult<T> result = new PageResult<>();
         result.setRecords(records);
         result.setTotal(total);
